requirement.txt
langgraph
langchain
langchain-google-genai
langchain-groq
langchain-tavily
deepagents
langchain-community
langchain_openai
ipykernel
ipywidgets
python-dotenv
biopython
langchain-qdrant
langchain-pinecone
pypdf
langchain_community